Use flying edges for mesh generation (#1741)
* adding-new-tables-for-flying-edges

* adding codes for flying-edges algo, still fixing left

* fixing laplace smoothing, only cleanups left


---------

+++ avogadro/rendering/meshgeometry.h
@@ -92,8 +92,8 @@ public:
    * the valid index range.
    * @{
    */
-  void addTriangle(unsigned int index1, unsigned int index2,
-                   unsigned int index3);
+  void addTriangle( size_t index1,  size_t index2,
+                    size_t index3);
   void addTriangles(const Core::Array<unsigned int>& indices);
   /** @} */
